:heavy_plus_sign: Removed code duplication by adding a new templaed function to create a component from a type and a list of args and a function to create a projectile.
